requirements.txt
boto3>=1.34.0
requests>=2.31.0
PyYAML>=6.0.1
